refactor: moves region.ts into utils/region.ts
